NewPollModal set up
