+++ views/poll-view/src/components/NewPollModal/index.ts
@@ -1,0 +1,3 @@
+import NewPollModal from "./NewPollModal";
+
+export default NewPollModal;
